+++ core-service/src/main/java/ru/geekbrains/webstore/core/entity/Order.java
@@ -1,5 +1,6 @@
 package ru.geekbrains.webstore.core.entity;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -10,8 +11,6 @@ import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import lombok.Data;
@@ -40,7 +39,7 @@ public class Order {
   private String address;
 
   @Column(name = "price")
-  private Long price;
+  private BigDecimal price;
 
   @OneToMany(mappedBy = "order", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
   private List<OrderItem> items = new ArrayList<>();
